Move EuroPrice to core package
