+++ src/main/java/com/ite/itea/ecommerce/domain/invoicing/LineItem.java
@@ -1,6 +1,6 @@
 package com.ite.itea.ecommerce.domain.invoicing;
 
-import com.ite.itea.ecommerce.domain.retail.EuroPrice;
+import com.ite.itea.ecommerce.domain.core.EuroPrice;
 
 /**
  * An invoiced item, i.e., a good or service at a certain price and quantity.
