+++ src/main/java/com/ite/itea/ecommerce/domain/retail/Picture.java
@@ -1,5 +1,7 @@
 package com.ite.itea.ecommerce.domain.retail;
 
+import com.ite.itea.ecommerce.domain.core.EuroPrice;
+
 public final class Picture extends Product {
 
     private final EuroPrice price;
