+++ src/main/java/com/ite/itea/ecommerce/domain/retail/ShoppingCart.java
@@ -1,5 +1,7 @@
 package com.ite.itea.ecommerce.domain.retail;
 
+import com.ite.itea.ecommerce.domain.core.EuroPrice;
+
 import java.util.ArrayList;
 import java.util.List;
