+++ src/main/java/com/ite/itea/ecommerce/domain/retail/Table.java
@@ -1,5 +1,7 @@
 package com.ite.itea.ecommerce.domain.retail;
 
+import com.ite.itea.ecommerce.domain.core.EuroPrice;
+
 public final class Table extends Product {
 
     private final EuroPrice legPrice;
