+++ src/main/java/com/ite/itea/ecommerce/domain/retail/Voucher.java
@@ -1,5 +1,7 @@
 package com.ite.itea.ecommerce.domain.retail;
 
+import com.ite.itea.ecommerce.domain.core.EuroPrice;
+
 public class Voucher {
 
     private final EuroPrice discountAmount;
